refactor: extract props and emits, define type FocusTrapInstance
