+++ packages/components/focus-trap/__tests__/focus-trap.test.ts
@@ -3,7 +3,7 @@ import { mount } from '@vue/test-utils'
 import { afterEach, describe, expect, it, vi } from 'vitest'
 import { EVENT_CODE } from '@element-plus/constants'
 import ElFocusTrap from '../src/focus-trap.vue'
-import { FOCUS_TRAP_INJECTION_KEY } from '../src/tokens'
+import { FOCUS_TRAP_INJECTION_KEY } from '../../../tokens/focus-trap'
 
 const AXIOM = 'rem is the best girl'
