+++ packages/components/focus-trap/index.ts
@@ -1,7 +1,7 @@
 import ElFocusTrap from './src/focus-trap.vue'
 
-export { ElFocusTrap }
-
 export default ElFocusTrap
-export * from './src/tokens'
+export * from '../../tokens/focus-trap'
 export * from './src/utils'
+
+export * from './src/focus-trap'
